refactor: add contextAPI
