+++ 레미_김민지/6주차 미션/실습-ToDoList ContextAPI/src/main.jsx
@@ -1,9 +1,12 @@
 import { StrictMode } from "react";
 import { createRoot } from "react-dom/client";
 import App from "./App.jsx";
+import { TodoContextProvider } from "./context/TodoContext.jsx";
 
 createRoot(document.getElementById("root")).render(
   <StrictMode>
-    <App />
+    <TodoContextProvider>
+      <App />
+    </TodoContextProvider>
   </StrictMode>,
 );
